ytAPI/requirements.txt
Django==3.1.4
django_filter==2.4.0
djangorestframework==3.13.1
google_api_python_client==2.33.0
